добавил скрипт инициализации таблицы комплекс мелкие правки

+++ src/main/resources/db/migration/V1__initialize_user.sql
@@ -3,7 +3,7 @@ SET FOREIGN_KEY_CHECKS = 0;
 DROP TABLE IF EXISTS users;
 
 CREATE TABLE users (
-  id                    INT(11) NOT NULL AUTO_INCREMENT,
+  id                    bigint NOT NULL AUTO_INCREMENT,
   username              VARCHAR(50) NOT NULL,
   password              CHAR(80) NOT NULL,
    PRIMARY KEY (id)
